fix: compile warnings in migrated infra headers
- dramtile_store.h: remove dead unused variable (dt_put_addr, dt_put)
- dramtile_store.h: fix 4UL→4ULL overflow (MinGW64 unsigned long=32bit)
- dramtile_store.h: suppress unused params in dt_putv stub
- dramtile_container.h: %zu→(unsigned)%u for MinGW printf compat

+++ core/infra/dramtile_container.h
@@ -158,8 +158,8 @@ static inline void dtc_print(DtContainer *c, const char *label) {
             dtc_is_delta(c) ? "|DELTA" : "");
     for (int i = 0; i < c->view.ndim; i++)
         fprintf(stderr, "%s%u", i ? "," : "", c->view.shape[i]);
-    fprintf(stderr, "] elem_sz=%zu strides=[", c->elem_size);
+    fprintf(stderr, "] elem_sz=%u strides=[", (unsigned)c->elem_size);
     for (int i = 0; i < c->view.ndim; i++)
-        fprintf(stderr, "%s%zu", i ? "," : "", c->stride[i]);
+        fprintf(stderr, "%s%u", i ? "," : "", (unsigned)c->stride[i]);
     fprintf(stderr, "] ptr=%p\n", (void*)c->view.data);
 }
